final attempt
What I had lying around after before I stopped working on this.

The main issue I was trying to solve for a bit was how to implement the
advanced filtering of GTG.

Using nested GtkFilterListModel s is a pain to manage and can't really
implement the bidirectional filtering of liblarch.

So I thought to create a thing that takes a GTG BaseTree and filters it
in a way similar to liblarch, but that was quite complicated and that
would be basically reinventing liblarch.

So instead I tried to create a BaseTree -> Liblarch translation layer,
then do the filtering in liblarch, and then reconvert it back to
BaseTree with another translation layer, which seemed to almost work,
but it randomly would segfault.